move coordinate as subpackage of model
