+++ single-line-diagram-core/src/main/java/com/powsybl/sld/layout/AbstractLayout.java
@@ -7,8 +7,8 @@
 package com.powsybl.sld.layout;
 
 import com.powsybl.commons.PowsyblException;
-import com.powsybl.sld.coordinate.Point;
 import com.powsybl.sld.model.*;
+import com.powsybl.sld.model.coordinate.Point;
 
 import java.util.*;
