+++ single-line-diagram-core/src/main/java/com/powsybl/sld/layout/AbstractVoltageLevelLayout.java
@@ -7,7 +7,7 @@
 package com.powsybl.sld.layout;
 
 import com.powsybl.sld.model.VoltageLevelGraph;
-import com.powsybl.sld.coordinate.Point;
+import com.powsybl.sld.model.coordinate.Point;
 import com.powsybl.sld.model.Node;
 
 import java.util.List;
