+++ single-line-diagram-core/src/main/java/com/powsybl/sld/layout/BlockPositionner.java
@@ -6,12 +6,12 @@
  */
 package com.powsybl.sld.layout;
 
-import com.powsybl.sld.coordinate.Position;
-import com.powsybl.sld.coordinate.Side;
 import com.powsybl.sld.model.*;
 import com.powsybl.sld.model.BusCell.Direction;
+import com.powsybl.sld.model.coordinate.Position;
+import com.powsybl.sld.model.coordinate.Side;
 
-import static com.powsybl.sld.coordinate.Position.Dimension.*;
+import static com.powsybl.sld.model.coordinate.Position.Dimension.*;
 
 import java.util.*;
 import java.util.stream.Collectors;
