+++ single-line-diagram-core/src/main/java/com/powsybl/sld/layout/HorizontalSubstationLayout.java
@@ -6,8 +6,8 @@
  */
 package com.powsybl.sld.layout;
 
-import com.powsybl.sld.coordinate.Point;
 import com.powsybl.sld.model.*;
+import com.powsybl.sld.model.coordinate.Point;
 
 import java.util.List;
