+++ single-line-diagram-core/src/main/java/com/powsybl/sld/layout/InfosNbSnakeLinesVertical.java
@@ -6,10 +6,10 @@
  */
 package com.powsybl.sld.layout;
 
-import com.powsybl.sld.coordinate.Side;
 import com.powsybl.sld.model.BusCell;
 import com.powsybl.sld.model.SubstationGraph;
 import com.powsybl.sld.model.VoltageLevelGraph;
+import com.powsybl.sld.model.coordinate.Side;
 
 import java.util.*;
 import java.util.function.Function;
