+++ single-line-diagram-core/src/main/java/com/powsybl/sld/layout/InternCellSide.java
@@ -6,8 +6,8 @@
  */
 package com.powsybl.sld.layout;
 
-import com.powsybl.sld.coordinate.Side;
 import com.powsybl.sld.model.InternCell;
+import com.powsybl.sld.model.coordinate.Side;
 
 import java.util.*;
