+++ single-line-diagram-core/src/main/java/com/powsybl/sld/layout/LegBusSet.java
@@ -6,8 +6,9 @@
  */
 package com.powsybl.sld.layout;
 
-import com.powsybl.sld.coordinate.Side;
 import com.powsybl.sld.model.*;
+import com.powsybl.sld.model.coordinate.Side;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
