+++ single-line-diagram-core/src/main/java/com/powsybl/sld/layout/Subsection.java
@@ -6,11 +6,11 @@
  */
 package com.powsybl.sld.layout;
 
-import com.powsybl.sld.coordinate.Side;
 import com.powsybl.sld.model.*;
+import com.powsybl.sld.model.coordinate.Side;
 
-import static com.powsybl.sld.coordinate.Side.LEFT;
-import static com.powsybl.sld.coordinate.Side.RIGHT;
+import static com.powsybl.sld.model.coordinate.Side.LEFT;
+import static com.powsybl.sld.model.coordinate.Side.RIGHT;
 
 import java.util.*;
 import java.util.function.Function;
