+++ single-line-diagram-core/src/main/java/com/powsybl/sld/layout/positionbyclustering/LBSClusterSide.java
@@ -6,11 +6,11 @@
  */
 package com.powsybl.sld.layout.positionbyclustering;
 
-import com.powsybl.sld.coordinate.Side;
 import com.powsybl.sld.layout.HorizontalBusLane;
 import com.powsybl.sld.layout.LBSCluster;
 import com.powsybl.sld.layout.LegBusSet;
 import com.powsybl.sld.model.*;
+import com.powsybl.sld.model.coordinate.Side;
 
 import java.util.*;
 import java.util.stream.Collectors;
