+++ single-line-diagram-core/src/main/java/com/powsybl/sld/layout/positionbyclustering/Links.java
@@ -6,9 +6,9 @@
  */
 package com.powsybl.sld.layout.positionbyclustering;
 
-import com.powsybl.sld.coordinate.Side;
 import com.powsybl.sld.layout.HorizontalBusLaneManager;
 import com.powsybl.sld.layout.LBSCluster;
+import com.powsybl.sld.model.coordinate.Side;
 
 import java.util.*;
