+++ single-line-diagram-core/src/main/java/com/powsybl/sld/library/AnchorPoint.java
@@ -8,7 +8,7 @@ package com.powsybl.sld.library;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.powsybl.sld.coordinate.Point;
+import com.powsybl.sld.model.coordinate.Point;
 
 import java.util.Objects;
