+++ single-line-diagram-core/src/main/java/com/powsybl/sld/model/AbstractParallelBlock.java
@@ -12,9 +12,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
-import static com.powsybl.sld.coordinate.Coord.Dimension.*;
-import static com.powsybl.sld.coordinate.Position.Dimension.*;
 import static com.powsybl.sld.model.Block.Extremity.*;
+import static com.powsybl.sld.model.coordinate.Coord.Dimension.*;
+import static com.powsybl.sld.model.coordinate.Position.Dimension.*;
 
 /**
  * @author Benoit Jeanson <benoit.jeanson at rte-france.com>
