+++ single-line-diagram-core/src/main/java/com/powsybl/sld/model/Block.java
@@ -7,10 +7,10 @@
 package com.powsybl.sld.model;
 
 import com.fasterxml.jackson.core.JsonGenerator;
-import com.powsybl.sld.coordinate.Coord;
-import com.powsybl.sld.coordinate.Orientation;
-import com.powsybl.sld.coordinate.Position;
 import com.powsybl.sld.layout.LayoutParameters;
+import com.powsybl.sld.model.coordinate.Coord;
+import com.powsybl.sld.model.coordinate.Orientation;
+import com.powsybl.sld.model.coordinate.Position;
 
 import java.io.IOException;
 import java.util.List;
