+++ single-line-diagram-core/src/main/java/com/powsybl/sld/model/BusCell.java
@@ -8,7 +8,7 @@ package com.powsybl.sld.model;
 
 import java.util.*;
 
-import com.powsybl.sld.coordinate.Orientation;
+import com.powsybl.sld.model.coordinate.Orientation;
 
 /**
  * @author Benoit Jeanson <benoit.jeanson at rte-france.com>
