+++ single-line-diagram-core/src/main/java/com/powsybl/sld/model/Edge.java
@@ -7,7 +7,7 @@
 package com.powsybl.sld.model;
 
 import com.fasterxml.jackson.core.JsonGenerator;
-import com.powsybl.sld.coordinate.Point;
+import com.powsybl.sld.model.coordinate.Point;
 
 import java.io.IOException;
 import java.util.Arrays;
