+++ single-line-diagram-core/src/main/java/com/powsybl/sld/model/FeederNode.java
@@ -8,7 +8,7 @@ package com.powsybl.sld.model;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.powsybl.commons.PowsyblException;
-import com.powsybl.sld.coordinate.Orientation;
+import com.powsybl.sld.model.coordinate.Orientation;
 
 import java.io.IOException;
 import java.util.Objects;
