+++ single-line-diagram-core/src/main/java/com/powsybl/sld/model/coordinate/Side.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.sld.coordinate;
+package com.powsybl.sld.model.coordinate;
 
 /**
  * @author Benoit Jeanson <benoit.jeanson at rte-france.com>
