+++ single-line-diagram-core/src/main/java/com/powsybl/sld/svg/DefaultSVGWriter.java
@@ -6,10 +6,10 @@
  */
 package com.powsybl.sld.svg;
 
-import com.powsybl.sld.coordinate.Point;
 import com.powsybl.sld.layout.LayoutParameters;
 import com.powsybl.sld.library.*;
 import com.powsybl.sld.model.Node;
+import com.powsybl.sld.model.coordinate.Point;
 import com.powsybl.sld.model.*;
 import com.powsybl.sld.svg.DiagramLabelProvider.Direction;
 import com.powsybl.sld.svg.GraphMetadata.FeederInfoMetadata;
@@ -32,8 +32,8 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import java.util.stream.Stream;
 
-import static com.powsybl.sld.coordinate.Position.Dimension.H;
 import static com.powsybl.sld.library.ComponentTypeName.*;
+import static com.powsybl.sld.model.coordinate.Position.Dimension.H;
 import static com.powsybl.sld.svg.DiagramStyles.*;
 
 /**
