+++ single-line-diagram-core/src/main/java/com/powsybl/sld/svg/WireConnection.java
@@ -6,11 +6,11 @@
  */
 package com.powsybl.sld.svg;
 
-import com.powsybl.sld.coordinate.Point;
 import com.powsybl.sld.library.AnchorOrientation;
 import com.powsybl.sld.library.AnchorPoint;
 import com.powsybl.sld.library.AnchorPointProvider;
 import com.powsybl.sld.model.BaseNode;
+import com.powsybl.sld.model.coordinate.Point;
 
 import java.util.ArrayList;
 import java.util.Collections;
