+++ single-line-diagram-core/src/test/java/com/powsybl/sld/iidm/TestSVGWriter.java
@@ -8,11 +8,11 @@ package com.powsybl.sld.iidm;
 
 import com.powsybl.iidm.network.Branch.Side;
 import com.powsybl.iidm.network.Network;
-import com.powsybl.sld.coordinate.Point;
-import com.powsybl.sld.coordinate.Position;
 import com.powsybl.sld.layout.LayoutParameters;
 import com.powsybl.sld.library.ComponentTypeName;
 import com.powsybl.sld.model.*;
+import com.powsybl.sld.model.coordinate.Point;
+import com.powsybl.sld.model.coordinate.Position;
 import com.powsybl.sld.svg.*;
 import org.junit.Before;
 import org.junit.Test;
